third_party/blink/web_tests/external/wpt/webrtc-encoded-transform/script-transform-generateKeyFrame.https-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught TypeError: Cannot read properties of undefined (reading 'getReader')
[FAIL] generateKeyFrame() throws for audio
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame(null) resolves for video sender, and throws for video receiver
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame throws NotAllowedError for invalid rid
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame throws NotFoundError for unknown rid
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame throws for unset transforms
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame timestamp should advance
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] await generateKeyFrame, await generateKeyFrame should see an increase in count of keyframes
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame rejects when the sender is negotiated inactive, and resumes succeeding when negotiated back to active
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame rejects when the sender is stopped, even without negotiation
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame rejects with a null track
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
Harness: the test ran to completion.
